+++ packages/devtools/client/src/entries/client/routes/react/[tab]/layout.config.tsx
@@ -1,0 +1,17 @@
+import { SelectLink } from '@/components/SelectLink';
+
+export const handle = {
+  breadcrumb: [
+    { title: 'React' },
+    {
+      title: (
+        <SelectLink
+          items={[
+            { to: '/react/components', title: 'Components' },
+            { to: '/react/profiler', title: 'Profiler' },
+          ]}
+        />
+      ),
+    },
+  ],
+};
