+++ packages/devtools/client/src/entries/client/routes/react/[tab]/layout.tsx
@@ -1,0 +1,5 @@
+import { Outlet } from '@modern-js/runtime/router';
+
+export default function Layout() {
+  return <Outlet />;
+}
