+++ packages/devtools/client/src/entries/client/routes/react/[tab]/page.module.scss
