+++ packages/devtools/client/src/entries/client/routes/react/page.loader.ts
@@ -1,0 +1,7 @@
+import { LoaderFunctionArgs, redirect } from '@modern-js/runtime/router';
+import { resolveURL } from 'ufo';
+
+export default ({ request }: LoaderFunctionArgs) => {
+  const target = resolveURL(request.url, './components');
+  return redirect(target);
+};
